updating configs
